feat: subscription tooltip
