+++ components/UI/styledTooltip.tsx
@@ -1,0 +1,14 @@
+import { Tooltip, TooltipProps, styled, tooltipClasses } from "@mui/material";
+
+const StyledToolTip = styled(({ className, ...props }: TooltipProps) => (
+  <Tooltip {...props} arrow classes={{ popper: className }} />
+))(() => ({
+  [`& .${tooltipClasses.arrow}`]: {
+    color: "#454545",
+  },
+  [`& .${tooltipClasses.tooltip}`]: {
+    backgroundColor: "#454545",
+  },
+}));
+
+export default StyledToolTip;
